rename: dir의 Prayer를 prayer로 변경 #21
